requirements.txt
rdflib>=7.0.0
SPARQLWrapper>=2.0.0
networkx>=3.0
matplotlib>=3.5.0
requests>=2.28.0
tqdm>=4.64.0
psutil>=5.9.0
pyyaml>=6.0
click>=8.0.0
retrying>=1.3.4
ftputil>=5.0.4 